[FE] mesh 구조 훅, 커스텀 훅 분리 및 SFU 로직 수정 (#396)
* refactor: hook 함수를 분리합니다.

* refactor: sfu 로직 수정
